fix compilation

+++ test/helpers/aggoracle_e2e.go
@@ -114,7 +114,7 @@ func CommonSetup(t *testing.T) (
 	// Syncer
 	dbPathSyncer := path.Join(t.TempDir(), "file::memory:?cache=shared")
 	maxRetries := 3
-	syncBlockChunkSize := 10
+	syncBlockChunkSize := uint64(10)
 	syncer, err := l1infotreesync.New(ctx, dbPathSyncer, gerL1Addr, common.Address{}, syncBlockChunkSize, etherman.LatestBlock, reorg, l1Client.Client(), time.Millisecond, 0, 100*time.Millisecond, maxRetries)
 	require.NoError(t, err)
 	go syncer.Start(ctx)
